pyproject.toml
[project]
name = "caliandra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.6",
    "django-cotton>=2.1.3",
    "pillow>=11.3.0",
]
